% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{FbarCards}
\alias{FbarCards}
\title{FbarCards Shiny App}
\usage{
FbarCards()
}
\value{
This function launches a shiny app and does not return a value.
}
\description{
Launches a 'shiny' app for the FbarCards game. In this game, a grid of cards
is displayed and the objective is to reorder the cards in each row such that,
when the rows are stacked, the columns of cards are in increasing order from
left to right. Players can swap the positions of two cards in the same row
before finalizing their choices and scoring the game. The game utilizes
Informative Hypothesis Testing (IHT) to score the final grid of cards.
}
\examples{
if (interactive()) {
  FbarCards()
}

}
