% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replext_helpers.R
\name{getCellBlocks}
\alias{getCellBlocks}
\title{Retrieve Cell Blocks for Replext Simulation}
\usage{
getCellBlocks()
}
\value{
A named list where each name corresponds to a user-friendly description
of a simulation type and each value is the function name that executes the
simulation.
}
\description{
This internal helper function provides a list of simulation types available in
the Replext Shiny application. It maps the user-friendly names of the
simulations to their corresponding function names.
}
\details{
The function currently supports 'Order Constrained ANOVA' and
'Inequality Constrained Linear Regression' simulations, corresponding to
\code{replext_t1_c1} and \code{replext_t2_c1} functions, respectively. This function is
used to populate the selection options in the Shiny application's UI.
}
\keyword{internal}
