% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replext_funcs.R
\name{mui}
\alias{mui}
\title{Calculate Group Means for One-Way ANOVA}
\usage{
mui(k, f)
}
\arguments{
\item{k}{An integer representing the number of groups (k = 3, ..., 8).}

\item{f}{A numeric value representing the effect size (f = 0.10, 0.15, 0.20,
0.25, 0.30, 0.40).Typical values represent small (0.10), medium (0.25), and
large (0.40) effects.}
}
\value{
A numeric vector containing the means of the groups. Each element of
the vector corresponds to a group mean.
}
\description{
This function calculates the means of different groups in a one-way ANOVA
setting. It uses the equally spaced difference calculated by 'd_eq_spaced' and
follows the approach described in Vanbrabant et al. (2015).
}
\examples{
mui(4, 0.25) # For k = 4 and f = 0.25
}
