% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nl2br}
\alias{nl2br}
\title{Convert Newline Characters to HTML Line Breaks}
\usage{
nl2br(x)
}
\arguments{
\item{x}{A character vector where each element is a string in which newline
characters should be replaced with \verb{<br/>} tags. The function vectorizes
over the character vector, replacing newline characters in each string.}
}
\value{
A character vector of the same length as \code{x}, where each string has
newline characters replaced with \verb{<br/>} tags.
}
\description{
This function takes a string and replaces all newline characters (\verb{\\n}) with
HTML line break tags (\verb{<br/>}), which can be useful when you want to render
text with preserved line breaks in a web interface, such as a Shiny app.
}
\keyword{internal}
