% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{str2list}
\alias{str2list}
\title{Convert a Character String to a List}
\usage{
str2list(arg_str)
}
\arguments{
\item{arg_str}{A character string representing R arguments.}
}
\value{
A list containing the arguments represented by \code{arg_str}. If
\code{arg_str} is not a valid representation of R arguments, the function
will throw an error.
}
\description{
This internal function converts a character string representing R arguments
into a list of arguments. It is primarily used to facilitate the passing of
additional arguments from the Shiny UI to internal functions within the app.
}
\keyword{internal}
