% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmir_model.R
\name{mmir_model}
\alias{mmir_model}
\title{Fit Restriktor Supported Model}
\usage{
mmir_model(formula, data, engine = "lm", standardize = TRUE, ...)
}
\arguments{
\item{formula}{An object of class \code{formula} (or one that can be coerced to
that class): a symbolic description of the model to be fitted.}

\item{data}{A data frame containing the variables in the model.}

\item{engine}{A character string indicating which engine to use for model
fitting.
Can be one of "lm", "glm", "rlm". Default is "lm".}

\item{standardize}{Logical. If TRUE, numeric variables in \code{data} are
standardized before fitting the model. Default is TRUE.}

\item{...}{Additional arguments to be passed to the model fitting function
(lm, glm, or rlm).}
}
\value{
An object representing the fitted model, of class 'lm', 'glm', or 'rlm'
depending on the engine used.
}
\description{
\code{mmir_model} is an internal function for the \code{mmirestriktor} package that fits
a model to data using one of the specified engines ('lm', 'glm', or 'rlm').
It also provides an option to standardize numeric variables.
}
\details{
The \code{mmir_model} function is a helper function for the \code{mmirestriktor} shiny
app. It allows users to fit models using one of the specified engines and
provides an option to standardize numeric variables.
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[stats]{glm}}, \code{\link[MASS]{rlm}}
}
\keyword{internal}
